Refactoring for thread safety
